+++ src/Events/UntranslatedMessage.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Translation\Events;
+
+final class UntranslatedMessage
+{
+    public function __construct(public readonly string $message)
+    {
+    }
+}