+++ tests/Nexendrie/Translation/Loaders/FileLoaderTestAbstract.php
@@ -3,6 +3,12 @@ declare(strict_types=1);
 
 namespace Nexendrie\Translation\Loaders;
 
+use Circli\EventDispatcher\EventDispatcher;
+use Circli\EventDispatcher\ListenerProvider\DefaultProvider;
+use Nexendrie\Translation\Events\FoldersChanged;
+use Nexendrie\Translation\Events\LanguageChanged;
+use Nexendrie\Translation\Events\LanguageLoaded;
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Tester\Assert;
 use Nexendrie\Translation\InvalidFolderException;
 use Nexendrie\Translation\FolderNotSetException;
@@ -15,6 +21,28 @@ use Nexendrie\Translation\FolderNotSetException;
 abstract class FileLoaderTestAbstract extends \Tester\TestCase
 {
     protected FileLoader $loader;
+    protected EventDispatcherInterface $eventDispatcher;
+    protected DefaultProvider $listenerProvider;
+    /**
+     * @var LanguageChanged[]|FoldersChanged[]|LanguageLoaded[]
+     */
+    private array $events = [];
+
+    protected function setUp(): void
+    {
+        $this->events = [];
+        $this->listenerProvider = new DefaultProvider();
+        $this->listenerProvider->listen(LanguageChanged::class, function (LanguageChanged $event): void {
+            $this->events[] = $event;
+        });
+        $this->listenerProvider->listen(FoldersChanged::class, function (FoldersChanged $event): void {
+            $this->events[] = $event;
+        });
+        $this->listenerProvider->listen(LanguageLoaded::class, function (LanguageLoaded $event): void {
+            $this->events[] = $event;
+        });
+        $this->eventDispatcher = new EventDispatcher($this->listenerProvider);
+    }
 
     public function testGetLang(): void
     {
@@ -25,10 +53,16 @@ abstract class FileLoaderTestAbstract extends \Tester\TestCase
 
     public function testSetLang(): void
     {
+        Assert::count(1, $this->events);
         $this->loader->lang = "cs";
         $lang = $this->loader->lang;
-        Assert::type("string", $lang);
         Assert::same("cs", $lang);
+        Assert::count(2, $this->events);
+        /** @var LanguageChanged $event */
+        $event = $this->events[1];
+        Assert::type(LanguageChanged::class, $event);
+        Assert::same("en", $event->oldLanguage);
+        Assert::same("cs", $event->newLanguage);
     }
 
     public function testGetFolders(): void
@@ -86,6 +120,7 @@ abstract class FileLoaderTestAbstract extends \Tester\TestCase
 
     public function testGetTexts(): void
     {
+        Assert::count(1, $this->events);
         $texts = $this->loader->getTexts();
         Assert::type("array", $texts);
         Assert::count(3, $texts);
@@ -93,6 +128,14 @@ abstract class FileLoaderTestAbstract extends \Tester\TestCase
         Assert::count(3, $texts["messages"]);
         Assert::type("array", $texts["book"]);
         Assert::count(5, $texts["book"]);
+        if (!$this instanceof MessagesCatalogueTest) {
+            Assert::count(2, $this->events);
+            /** @var LanguageLoaded $event */
+            $event = $this->events[1];
+            Assert::type(LanguageLoaded::class, $event);
+            Assert::same("en", $event->language);
+        }
+
         $this->loader->lang = "cs";
         $texts = $this->loader->getTexts();
         Assert::type("array", $texts);
@@ -104,6 +147,17 @@ abstract class FileLoaderTestAbstract extends \Tester\TestCase
         if ($this->loader instanceof MessagesCatalogue) {
             return; // the following tests for some reason fail with MessagesCatalogue
         }
+        Assert::count(4, $this->events);
+        /** @var LanguageChanged $event */
+        $event = $this->events[2];
+        Assert::type(LanguageChanged::class, $event);
+        Assert::same("en", $event->oldLanguage);
+        Assert::same("cs", $event->newLanguage);
+        /** @var LanguageLoaded $event */
+        $event = $this->events[3];
+        Assert::type(LanguageLoaded::class, $event);
+        Assert::same("cs", $event->language);
+
         $this->loader->lang = "xyz";
         $texts = $this->loader->getTexts();
         Assert::type("array", $texts);
@@ -112,6 +166,16 @@ abstract class FileLoaderTestAbstract extends \Tester\TestCase
         Assert::count(3, $texts["messages"]);
         Assert::type("array", $texts["book"]);
         Assert::count(5, $texts["book"]);
+        Assert::count(6, $this->events);
+        /** @var LanguageChanged $event */
+        $event = $this->events[4];
+        Assert::type(LanguageChanged::class, $event);
+        Assert::same("cs", $event->oldLanguage);
+        Assert::same("xyz", $event->newLanguage);
+        /** @var LanguageLoaded $event */
+        $event = $this->events[5];
+        Assert::type(LanguageLoaded::class, $event);
+        Assert::same("xyz", $event->language);
     }
 
     public function testNoFolder(): void